fix: add Evented mixin to ExtensionManager for trigger support
Error Fixed:
TypeError: this.trigger is not a function

The ExtensionManager uses this.trigger('engine.loaded', engineInstance)
to emit events when engines are loaded, but the Service class doesn't
include the Evented mixin by default.

Changes:
- Import Evented from '@ember/object/evented'
- Extend Service with Evented mixin: Service.extend(Evented)

This allows the service to:
- trigger('event', ...args) - emit events
- on('event', callback) - listen to events
- off('event', callback) - remove listeners

Useful for components/services that need to react to engine loading.